+++ src/converse-notification.js
@@ -32,6 +32,7 @@
             this.updateSettings({
                 notify_all_room_messages: false,
                 show_desktop_notifications: true,
+                show_chatstate_notifications: false,
                 chatstate_notification_blacklist: [],
                 // ^ a list of JIDs to ignore concerning chat state notifications
                 play_sounds: false,
@@ -211,7 +212,7 @@
                  * Will show an HTML5 notification to indicate that the chat
                  * status has changed.
                  */
-                if (_converse.areDesktopNotificationsEnabled()) {
+                if (_converse.areDesktopNotificationsEnabled() && _converse.show_chatstate_notifications) {
                     _converse.showChatStateNotification(contact);
                 }
             };